Proposed fix for error when downloading live younow streams to directories with spaces
I've identified a problem when recording a stream to a directory that has spaces (happened when running the auto downloader script), and propose this fix for it. Not sure if it happens elsewhere in the code.

I've done this edit on my machine and it works, but i'm not familiar with the .ts format, so i'm unsure if i should edit that file too with the fix, changed only the .js.